Fix symbol renaming for cross-compiled targets
The symbols generated for the library is renamed with a prefix
"psa_crypto_0_11_0_" to avoid any collisions. This is done using
objcopy tool. The same native tool is used for both native builds
and cross compilation but the tool fails to rename the symbols for
cross compiled library throwing the following error:

"objcopy: Unable to recognise the format of the input file"

The exit status is 0 when this happens and build.rs does not
identify the issue due to this. The generated libraries still have
the old symbols and this is reflected only when it is linked.

The CI run for parsec cross compilation highlighted this bug.
https://github.com/parallaxsecond/parsec/actions/runs/6480791927/job/17597014985?pr=718

The current patch uses the objcopy tool from the cross compiler
toolchain to ensure proper symbols are generated.

Signed-off-by: Gowtham Suresh Kumar <gowtham.sureshkumar@arm.com>